Improve error handling when placing messages on dead letter queue
If there's an error trying to put a message on the dead letter queue,
attempt to put just the message headers on the queue. (This should
handle the case where the root cause of the error is that the message
is too large for SQS queues.) If that fails, just log the message,
but don't add to a queue.